chore: refactor Growl component files
